+++ src/Telegram.Bot/Types/FileBase.cs
@@ -7,12 +7,18 @@ namespace Telegram.Bot.Types
     /// This object represents a file ready to be downloaded. The file can be downloaded via <see cref="TelegramBotClient.GetFileAsync"/>. It is guaranteed that the link will be valid for at least 1 hour. When the link expires, a new one can be requested by calling <see cref="TelegramBotClient.GetFileAsync"/>.
     /// </summary>
     [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
-    public class File : FileBase
+    public abstract class FileBase
     {
         /// <summary>
-        /// File path. Use <see cref="TelegramBotClient.GetFileAsync"/> to get the file.
+        /// Unique identifier for this file
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public string FileId { get; set; }
+
+        /// <summary>
+        /// Optional. File size, if known
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string FilePath { get; set; }
+        public int FileSize { get; set; }
     }
 }