+++ test/Telegram.Bot.Tests.Integ/Sending Messages/SendingVideoMessageTests.cs
@@ -83,7 +83,6 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
             Assert.Equal(duration, message.VideoNote.Duration);
             Assert.Equal(widthAndHeight, message.VideoNote.Length);
             Assert.NotEmpty(message.VideoNote.Thumb.FileId);
-            Assert.NotEmpty(message.VideoNote.Thumb.FilePath);
             Assert.True(message.VideoNote.Thumb.FileSize > 200);
             Assert.True(message.VideoNote.Thumb.Width > 50);
             Assert.True(message.VideoNote.Thumb.Height > 50);